update BasMove to use u8

+++ src/core/move.rs
@@ -10,13 +10,13 @@ include!("../util/generated_macro.rs");
 #[derive(Debug, PartialEq, Eq)]
 #[derive(Clone, Copy, Ord, PartialOrd)]
 pub struct BaseMove {
-    pub from: usize,
-    pub to: usize,
+    pub from: u8,
+    pub to: u8,
     pub capture: bool,
 }
 
 impl BaseMove {
-    pub(crate) fn new(from: usize, to: usize, capture: bool) -> BaseMove {
+    pub(crate) fn new(from: u8, to: u8, capture: bool) -> BaseMove {
         BaseMove { from, to, capture }
     }
 }
@@ -69,27 +69,27 @@ impl fmt::Display for Move {
 
     }
 }
-fn write_default(from: usize, to: usize, capture: bool) -> String {
-    format!("{}{}{}", format_square(from), if capture { 'x' } else { '-' }, format_square(to))
+fn write_default(from: u8, to: u8, capture: bool) -> String {
+    format!("{}{}{}", format_square(from as usize), if capture { 'x' } else { '-' }, format_square(to as usize))
 }
 
 #[derive(Debug, PartialEq, Eq)]
 #[derive(Clone, Copy)]
 pub struct RawMove {
-    pub from: usize,
-    pub to: usize,
+    pub from: u8,
+    pub to: u8,
     pub promote_to: Option<PieceType>,
 }
 
 impl RawMove {
-    pub(crate) fn new(from: usize, to: usize, promote_to: Option<PieceType>) -> RawMove {
+    pub(crate) fn new(from: u8, to: u8, promote_to: Option<PieceType>) -> RawMove {
         RawMove {from, to, promote_to}
     }
 }
 impl fmt::Display for RawMove {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
         let promote_to: String = self.promote_to.map_or(String::new(), |piece_type| piece_type.first_letter().to_lowercase().to_string());
-        write!(f, "{}{}{}", format_square(self.from), format_square(self.to), promote_to)
+        write!(f, "{}{}{}", format_square(self.from as usize), format_square(self.to as usize), promote_to)
     }
 }
